Fix unit tests for Arkane reading of P and T limits.
Because the T limits were being read incorrectly,
a test of the rate coefficient that was supposedly
at Tmax (but was in fact at 700K) was failing once
the T limit reading was fixed.
Also the pressure limit reading test was using .value
instead of .value_si so that broke also.