Allow comment parsing to not die from "in family __" comment.
It doesn't actually _parse_ the new comment to figure out what
family the thing comes from, but in most contexts it is not needed.
At least now it doesn't crash. 
Also added the new phrase to at least one item in the testing examples
so the unit tests could catch this.

Co-authored-by: Richard West <r.west@northeastern.edu>
Co-authored-by: Sevy Harris <sevy.harris@gmail.com>