Modify sanitization to accommodate kekulization
Some RMG molecule representations are not representable in RDKit.
This commit makes it so that if the kekulization step fails
during sanitization, then RDKit will try to sanitize
with less stringent sanitization criteria.

This was necessary because some species from adjacency lists
would pop up and crash out due to failed sanitization.