Possible simplification of to_rdkit_mol for cutting labels.
We do a call to detect_cutting_label (which is a complicated function
to find them buried inside SMILES strings) passing it the label of
every atom. (Though maybe these are mostly ''??).
Anyway, I'm hopeful we can just see if the atom is a CuttingLabel.
This commit puts in some Assert statements that should, if I'm right,
never fail.

If that's the case, we can then make the change.