Possible simplification of to_rdkit_mol for cutting labels.
We do a call to detect_cutting_label (which is a complicated function
to find them buried inside SMILES strings) passing it the label of
every atom. (Though maybe these are mostly ''?).

This commit checks whether a simpler approach would work: just
checking if the atom.label is either 'R' or 'L'.

If that's the case, we can then make the change.