Add test_deterministic_bicyclic_decomposition
Test that the decomposition of a polyring into bicyclics and then into single rings
is deterministic. This is important because the thermo estimation depends on the
order of the rings. Currently this is not guaranteed, so if this test fails, we
just skip it.

See https://github.com/ReactionMechanismGenerator/RMG-Py/issues/2562