Added completedNetworks option to prevent exploring PDep networks that are complete.
Sometimes you have a pressure-dependent network that has been studied 
in great detail and put in a seed mechanism or reaction library,
(such as CH2O2) and you don't want RMG to add reactions to the network
and your model, because doing so will make it worse.
Any reactions net reactions that are in your seed mechanism won't be changed,
but currently if there's a new reaction RMG would add it.

This option lets you specify in the input file that you don't want
RMG to expand certain networks further. You specify them with a line
in your pressureDependence block of your input file like

   completedNetworks = ['CH2O2', 'C2H6'],